Fix T43871: Selecting Boundary Loop While Holding Shift Fails
Fixes T43871
I guess everything needed to understand what's going on here can be read out of the comment I added in the patch.
It is quite hard to find a good solution for this as we'd need more modifier keys again to support this properly, but I think this is fine for now.