UI: Draw scroll-bars overlapping with editor content
Scroll-bars used to draw in a little extra space in the editor, causing
buttons to jump a bit when they appeared/disappeared. Now they draw on
top of the buttons, just big enough to avoid non-minor overlaps. This
can, and probably will get further adjustments.

With this we can get rid of a hack that was calling the (often Python
defined) panel definition - the panel 'draw' callback - twice.

The old code to draw scrollbars next to, instead of over the editor
contents is still there, just #ifdef'ed out. I will remove it once I'm
confident the new solution works well.